+++ test-curves/benches/bls12_381.rs
@@ -1,4 +1,4 @@
-use ark_algebra_bench_templates::*;
+use ark_algebra_bench_templates::{bench, criterion_main, field_common, paste, prime_field, sqrt};
 use ark_test_curves::bls12_381::{
     fq::Fq, fq2::Fq2, fr::Fr, Bls12_381, Fq12, G1Projective as G1, G2Projective as G2,
 };